Fix realm parameter in Authorization header
Made realm the first parameter and left value unencoded.

My specific use case requires the realm parameter to be first parameter in the Authorization header. In reviewing the spec, the realm parameter isn't normalized with other parameters and examples in OAuth protocol and other OAuth libraries seem to list realm as first parameter in the header.
http://tools.ietf.org/html/rfc5849#section-3.5.1

Examples in the OAuth spec don't appear to be encoded strings and the protocol states the value as only a quoted string.
http://tools.ietf.org/html/rfc2617#section-1.2